refactor: restructured files and folders for multiple build target support
