+++ package.json
@@ -3,18 +3,19 @@
   "version": "1.2.1",
   "description": "Cross browser Speech Synthesis",
   "type": "module",
-  "main": "index.js",
-  "browser": "index.js",
+  "main": "dist/EasySpeech.cjs.js",
+  "browser": "dist/EasySpeech.iife.js",
+  "exports": "dist/EasySpeech.js",
   "scripts": {
     "test": "NODE_ENV=test ./node_modules/.bin/mocha 'tests/**/*.tests.js'",
     "test-debug": "NODE_ENV=test ./node_modules/.bin/mocha  --inspect --debug-brk 'tests/**/*.tests.js'",
     "test:coverage": "NODE_ENV=test c8 --include=index.js --reporter=html --reporter=text ./node_modules/.bin/mocha 'tests/**/*.tests.js'",
     "lint": "standard . | snazzy",
     "lint:fix": "standard . --fix | snazzy",
-    "prepublish": "npm run lint && npm run test && npm run docs:api && npm run docs:html",
+    "prepublish": "npm run lint && npm run test && npm run build && npm run build:docs",
     "build:docs": "jsdoc2md index.js > API.md",
     "build:demo": "rollup --config rollup.demo.config.js",
-    "build": "rollup --config rollup.config.js"
+    "build": "rollup --config rollup.dist.config.js"
   },
   "repository": {
     "type": "git",
@@ -56,12 +57,16 @@
     "node": ">= 14.x"
   },
   "files": [
-    "index.js",
-    "API.md"
+    "dist",
+    "API.md",
+    "README.md",
+    "LICENSE",
+    "CHANGELOG.md"
   ],
   "standard": {
     "ignore": [
-      "docs/bundle.js"
+      "dist",
+      "docs"
     ]
   }
 }
