+++ rollup.dist.config.js
@@ -1,30 +1,36 @@
 import { babel, getBabelOutputPlugin } from '@rollup/plugin-babel'
 import commonjs from '@rollup/plugin-commonjs'
 
-export default {
-  input: 'index.js',
+export default [{
+  input: 'src/EasySpeech.js',
   output: [
     {
-      file: 'dist/bundle.es5.js',
+      file: 'dist/EasySpeech.js',
+      format: 'es'
+    },
+    {
+      file: 'dist/EasySpeech.es5.js',
       format: 'es',
       plugins: [getBabelOutputPlugin({ presets: ['@babel/preset-env'] })]
     },
     {
-      file: 'dist/bundle.cjs.js',
+      file: 'dist/EasySpeech.cjs.js',
       format: 'cjs',
       plugins: [getBabelOutputPlugin({ presets: ['@babel/preset-env'] })]
-    },
-    {
-      file: 'dist/bundle.iife.js',
-      format: 'iife',
-      name: 'EasySpeech'
     }
-  ],
+  ]
+}, {
+  input: 'src/EasySpeech.js',
+  output: {
+    file: 'dist/EasySpeech.iife.js',
+    format: 'iife',
+    name: 'EasySpeech'
+  },
   plugins: [
     commonjs(),
     babel({
       exclude: 'node_modules/**',
       babelHelpers: 'bundled'
-    }),
+    })
   ]
-}
+}]
